Bot Improvements
Added the Remaining attribute so strings can pull any remaining strings passed to it.

In addition, it will attempt to parse arguments passed to it to the types requested by the command method, and fail gracefully if it cannot.

More error checks and quality of life features were added to the bot class.