Command Pre/PostProcessors
Added ICommandProcessor with Preprocess and Postprocess members to allow action to be taken before and after a command is executed.

Any and everything permissions related was moved to Permissions.cs
PermissionsProcessor is now loaded as a default commandProcessor for Bot

I changed a few namespace names to avoid name conflicts.